+++ src/app/run/voxel_data_generator.rs
@@ -18,14 +18,14 @@ fn generate_tree_layer_rec(dst: &mut Vec<VoxelData>, layer_idx: u8, max_layer: u
     let data_index = dst.len() - 1;
 
     
-    dst[data_index].child_indicies[0] = generate_tree_layer_rec(dst, layer_idx + 1, max_layer).max(0) as u16;
-    dst[data_index].child_indicies[1] = generate_tree_layer_rec(dst, layer_idx + 1, max_layer).max(0) as u16;
-    dst[data_index].child_indicies[2] = generate_tree_layer_rec(dst, layer_idx + 1, max_layer).max(0) as u16;
-    dst[data_index].child_indicies[3] = generate_tree_layer_rec(dst, layer_idx + 1, max_layer).max(0) as u16;
-    dst[data_index].child_indicies[4] = generate_tree_layer_rec(dst, layer_idx + 1, max_layer).max(0) as u16;
-    dst[data_index].child_indicies[5] = generate_tree_layer_rec(dst, layer_idx + 1, max_layer).max(0) as u16;
-    dst[data_index].child_indicies[6] = generate_tree_layer_rec(dst, layer_idx + 1, max_layer).max(0) as u16;
-    dst[data_index].child_indicies[7] = generate_tree_layer_rec(dst, layer_idx + 1, max_layer).max(0) as u16;
+    dst[data_index].child_indicies[0] = generate_tree_layer_rec(dst, layer_idx + 1, max_layer).max(0) as u32;
+    dst[data_index].child_indicies[1] = generate_tree_layer_rec(dst, layer_idx + 1, max_layer).max(0) as u32;
+    dst[data_index].child_indicies[2] = generate_tree_layer_rec(dst, layer_idx + 1, max_layer).max(0) as u32;
+    dst[data_index].child_indicies[3] = generate_tree_layer_rec(dst, layer_idx + 1, max_layer).max(0) as u32;
+    dst[data_index].child_indicies[4] = generate_tree_layer_rec(dst, layer_idx + 1, max_layer).max(0) as u32;
+    dst[data_index].child_indicies[5] = generate_tree_layer_rec(dst, layer_idx + 1, max_layer).max(0) as u32;
+    dst[data_index].child_indicies[6] = generate_tree_layer_rec(dst, layer_idx + 1, max_layer).max(0) as u32;
+    dst[data_index].child_indicies[7] = generate_tree_layer_rec(dst, layer_idx + 1, max_layer).max(0) as u32;
 
 
     data_index as isize
